Add files via upload
Project Title: God of War Game Loading Page

Description:
This project is a loading page inspired by the popular game "God of War." Created as part of my web development internship at Octanet, the loading page serves as an engaging and visually appealing introduction to a gaming experience.

Utilizing HTML and CSS, I designed a dynamic background that captures the essence of the game, incorporating elements such as iconic symbols, characters, and themes from the God of War universe. Unlike traditional static loading pages, this dynamic design provides users with an immersive preview of what awaits them in the game.

Key Features:

Dynamic background: The background of the loading page changes dynamically, creating a visually captivating experience for users.
God of War theme: The design incorporates elements from the God of War game, including symbols, characters, and themes, to immerse users in the gaming atmosphere.
Smooth transitions: CSS animations and transitions are employed to ensure smooth loading transitions and enhance user engagement.
This project not only demonstrates my proficiency in HTML and CSS but also showcases my creativity and attention to detail in web design. It serves as an exciting introduction to the gaming experience, setting the stage for an immersive journey into the world of God of War